emotion-recognition-system/requirements.txt
deepface>=0.0.79
tf-keras>=2.15.0
flask>=3.0.0
flask-cors>=4.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
